requirements.txt
flask==3.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
langchain-core==0.1.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
duckduckgo-search==4.1.1
pydantic==2.5.0
pytesseract==0.3.10
pdf2image==1.16.3
pypdfium2==4.26.0
unstructured==0.11.6
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
pandas==2.0.3